removed AttributeDefinitionDraft
Signature is the same as AttributeDefinition.
The only difference lies in the name being required, but this isn't enforced in the SDK itself.
To simplify the interface we can just utilize the AttributeDefinition when creating a new attribute
and perform validations in the create/update methods themselve (or leave it as it is right now
and let the commercetool API return an error)