Refactor service generation
This commit is quite large and is the result of incremental fixes all
over the service generation codebase.

The result of this commit is that all resources (services) are now
correctly parsed and that the API should now cover the complete
specification.

This also fixes decoding time.Time objects which was throwing an error
while testing the GetMessageWithID() function (which is a special case in
itself since it returns an interface instead of a struct)